Parse the Author field from a DESCRIPTION file

Parse the Author field from a DESCRIPTION file into a person object

Usage

parse_authors(author_string)

Arguments

author_string

A character containing the Author or Maintainer field from a DESCRIPTION file

Value

A character vector, or a list of character vectors of length equals to the length of author_string

Examples

# Read from a DESCRIPTION file directly
utils_description <- system.file("DESCRIPTION", package = "utils")
utils_authors <- read.dcf(utils_description, "Author")

parse_authors(utils_authors)
#> [1] "R Core Team"            "contributors worldwide"

# Read from a database of CRAN metadata
cran_epidemiology_packages$Author |>
  parse_authors() |>
  unlist() |>
  unique() |>
  sort()
#>   [1] "A. Riebler"                                                                     
#>   [2] "Aaron A. King"                                                                  
#>   [3] "Abdoelnaser Degoot"                                                             
#>   [4] "Abdoelnaser M. Degoot"                                                          
#>   [5] "Adam J Kucharski"                                                               
#>   [6] "Adam Kucharski"                                                                 
#>   [7] "Adam Omidpanah"                                                                 
#>   [8] "Adrian Prodan"                                                                  
#>   [9] "Adrien Le Guillou"                                                              
#>  [10] "Ahmad Rabiee"                                                                   
#>  [11] "Ahmad family = Rabiee"                                                          
#>  [12] "Alan C. Hindmarsh"                                                              
#>  [13] "Albert Y. Kim"                                                                  
#>  [14] "Alex Richard Cook"                                                              
#>  [15] "Alex W. Friedrich"                                                              
#>  [16] "Alex Welte"                                                                     
#>  [17] "Alexander D. Becker"                                                            
#>  [18] "Alexander Kowarik"                                                              
#>  [19] "Alexander W. Friedrich"                                                         
#>  [20] "Alexandre Courtiol"                                                             
#>  [21] "Alexis Robert"                                                                  
#>  [22] "Alexis Vittengl"                                                                
#>  [23] "Ali Ezzat"                                                                      
#>  [24] "Allison Cheung"                                                                 
#>  [25] "Amanda Dobbyn"                                                                  
#>  [26] "Amit Sandhel"                                                                   
#>  [27] "Andrea Riebler"                                                                 
#>  [28] "Andreas Handel"                                                                 
#>  [29] "Andree Valle-Campos"                                                            
#>  [30] "Andrew Azman"                                                                   
#>  [31] "Andrew Johnson"                                                                 
#>  [32] "Andrew P. Norgan"                                                               
#>  [33] "Aniko Szabo"                                                                    
#>  [34] "Anita Williams"                                                                 
#>  [35] "Anne Cori"                                                                      
#>  [36] "Annick Lenglet"                                                                 
#>  [37] "Anthony Underwood"                                                              
#>  [38] "Anton Mymrikov"                                                                 
#>  [39] "Ariana Wijermans"                                                               
#>  [40] "Arthur Allignol"                                                                
#>  [41] "Attractive Chaos"                                                               
#>  [42] "Avery McIntosh"                                                                 
#>  [43] "Bankolé Ahadzie"                                                                
#>  [44] "Bart C. Meijer"                                                                 
#>  [45] "Ben Listyg"                                                                     
#>  [46] "Bendix Carstensen"                                                              
#>  [47] "Bernard Rosner"                                                                 
#>  [48] "Bertrand Sudre"                                                                 
#>  [49] "Bhanu N. M. Sinha"                                                              
#>  [50] "Bhanu Sinha"                                                                    
#>  [51] "Bin Li"                                                                         
#>  [52] "Bob Gray"                                                                       
#>  [53] "Brian J. Knaus"                                                                 
#>  [54] "Brian Knaus"                                                                    
#>  [55] "Brian McKay"                                                                    
#>  [56] "Bruce E. Kendall"                                                               
#>  [57] "Bruce Kendall"                                                                  
#>  [58] "Bubacarr Bah"                                                                   
#>  [59] "C. Lang"                                                                        
#>  [60] "Caitlin Collins"                                                                
#>  [61] "Cari Van Schalkwyk"                                                             
#>  [62] "Carl A. B. Pearson"                                                             
#>  [63] "Carles Breto"                                                                   
#>  [64] "Carles Bretó"                                                                   
#>  [65] "Carles Martinez Breto"                                                          
#>  [66] "Casper J. Albers"                                                               
#>  [67] "Chad Klumb"                                                                     
#>  [68] "Charles Reynard"                                                                
#>  [69] "Charlie Whittaker"                                                              
#>  [70] "Chase Golden"                                                                   
#>  [71] "Chris Groendyke"                                                                
#>  [72] "Chris Hartgerink"                                                               
#>  [73] "Christian E. Galarza"                                                           
#>  [74] "Christian F. Luz"                                                               
#>  [75] "Christian Lang"                                                                 
#>  [76] "Christine Casey"                                                                
#>  [77] "Christophe Fraser"                                                              
#>  [78] "Christopher Endres"                                                             
#>  [79] "Christopher J. Endres"                                                          
#>  [80] "Christopher Jarvis"                                                             
#>  [81] "Cici Chen"                                                                      
#>  [82] "Clement W. Ulrich"                                                              
#>  [83] "Cleve Moler"                                                                    
#>  [84] "Cody Dailey"                                                                    
#>  [85] "Cord Heuer"                                                                     
#>  [86] "Corinna Glasner"                                                                
#>  [87] "Cristiano Santos"                                                               
#>  [88] "D. Sabanés Bové"                                                                
#>  [89] "Damjan Manevski"                                                                
#>  [90] "Daniel C. Reuman"                                                               
#>  [91] "Daniel Maillard"                                                                
#>  [92] "Daniel May"                                                                     
#>  [93] "Daniel Pfeffer"                                                                 
#>  [94] "Daniel Sabanes Bove"                                                            
#>  [95] "Daniel Sabanés Bové"                                                            
#>  [96] "Daniel Wollschlaeger"                                                           
#>  [97] "Daniel Wollschlaeger User"                                                      
#>  [98] "Dao Nguyen"                                                                     
#>  [99] "David Ardia"                                                                    
#> [100] "David Clayton"                                                                  
#> [101] "David Hunter"                                                                   
#> [102] "David R. Hunter"                                                                
#> [103] "David Smith"                                                                    
#> [104] "David Welch"                                                                    
#> [105] "Deepayan Sarkar"                                                                
#> [106] "Denis Haine"                                                                    
#> [107] "Denis Laloe"                                                                    
#> [108] "Dennis Souverein"                                                               
#> [109] "Dirk Schumacher"                                                                
#> [110] "Dmytro Mykhailenko"                                                             
#> [111] "Dominique Pontier"                                                              
#> [112] "Dustin Tingley"                                                                 
#> [113] "E. S. Venkatraman"                                                              
#> [114] "Eamon B. O'Dea"                                                                 
#> [115] "Eamon O'Dea"                                                                    
#> [116] "Ed Ionides"                                                                     
#> [117] "Edouard F. Bonneville"                                                          
#> [118] "Eduard Grebe"                                                                   
#> [119] "Edward L. Ionides"                                                              
#> [120] "Edwin Van Leeuwen"                                                              
#> [121] "Elisabeth Dahlqwist"                                                            
#> [122] "Elizabeth Rogawski McQuade"                                                     
#> [123] "Emanuele Guidotti"                                                              
#> [124] "Emily Beylerian"                                                                
#> [125] "Emma Mendelsohn"                                                                
#> [126] "EpiForecasts"                                                                   
#> [127] "Eric Brown"                                                                     
#> [128] "Eric H. L. C. M. Hazenberg"                                                     
#> [129] "Ernst Hairer"                                                                   
#> [130] "Erwin E. A. Hassing"                                                            
#> [131] "Erwin E.A. Hassing"                                                             
#> [132] "Esa Laara"                                                                      
#> [133] "Esa Läärä"                                                                      
#> [134] "Esther Kissling"                                                                
#> [135] "European Union"                                                                 
#> [136] "Evan Sergeant"                                                                  
#> [137] "Fabio Demarqui"                                                                 
#> [138] "Federico Calboli"                                                               
#> [139] "Finlay Campbell"                                                                
#> [140] "Flavio Finger"                                                                  
#> [141] "Francisco Orchard"                                                              
#> [142] "Francois Balloux"                                                               
#> [143] "Frédéric Michaud"                                                               
#> [144] "Geoff Jones"                                                                    
#> [145] "George D. Byrne"                                                                
#> [146] "Gerry Tonkin-Hill"                                                              
#> [147] "Guy Baele"                                                                      
#> [148] "Gwen Knight"                                                                    
#> [149] "Haixiang Zhang"                                                                 
#> [150] "Hamada S. Badr"                                                                 
#> [151] "Hao Zhu"                                                                        
#> [152] "Harry Gibson"                                                                   
#> [153] "Hein Putter"                                                                    
#> [154] "Helen Wearing"                                                                  
#> [155] "Henok Woldu"                                                                    
#> [156] "Hilmarie Brand"                                                                 
#> [157] "Howard Burkom"                                                                  
#> [158] "Hugo Gruson"                                                                    
#> [159] "Hü¾\u008d¦\u0098¼hle"                                                           
#> [160] "Imperial College of Science"                                                    
#> [161] "Isaac Chun-Hai Fung"                                                            
#> [162] "Isaac Fung"                                                                     
#> [163] "Ismail Ahmed"                                                                   
#> [164] "Isobel Blake"                                                                   
#> [165] "Ivana Rodriguez Ewerlöf"                                                        
#> [166] "J Miettinen"                                                                    
#> [167] "J Pitkaniemi"                                                                   
#> [168] "Jake Stockwin"                                                                  
#> [169] "Jakob Schumacher"                                                               
#> [170] "James M. Azam"                                                                  
#> [171] "James Munday"                                                                   
#> [172] "Janetta Skarp"                                                                  
#> [173] "Jarrett Byrnes"                                                                 
#> [174] "Javier Sanchez"                                                                 
#> [175] "Jean Pierre Decorps"                                                            
#> [176] "Jeffrey Eaton"                                                                  
#> [177] "Jen Rozier"                                                                     
#> [178] "Jennifer F. Bobb"                                                               
#> [179] "Jeno Reiczigel"                                                                 
#> [180] "Jessica Grembi"                                                                 
#> [181] "Jim Robison-Cox"                                                                
#> [182] "Jing Ma"                                                                        
#> [183] "Jo van Schalkwyk"                                                               
#> [184] "Joana Gomes Dias"                                                               
#> [185] "Joe Hickson"                                                                    
#> [186] "Joel Hellewell"                                                                 
#> [187] "Johann Popp"                                                                    
#> [188] "John Rossow"                                                                    
#> [189] "Jon Wakefield"                                                                  
#> [190] "Jonas Salm"                                                                     
#> [191] "Jonathan A. Polonsky"                                                           
#> [192] "Jonathan M. Olate"                                                              
#> [193] "Jonathon Marshall"                                                              
#> [194] "Joonas Miettinen"                                                               
#> [195] "Jorge Chavarro"                                                                 
#> [196] "Jose E. Lozano"                                                                 
#> [197] "Jose E. Lozano Alonso"                                                          
#> [198] "Joshua W. Lambert"                                                              
#> [199] "Judith M. Fonville"                                                             
#> [200] "Juliane Manitz"                                                                 
#> [201] "Juliet R.C. Pulliam"                                                            
#> [202] "Jun Cai"                                                                        
#> [203] "Justin Lessler"                                                                 
#> [204] "K Seppa"                                                                        
#> [205] "Karim Mané"                                                                     
#> [206] "Karline Soetaert"                                                               
#> [207] "Karri Seppa"                                                                    
#> [208] "Katayoun Moazami-Goudarzi"                                                      
#> [209] "Katelyn Gostic"                                                                 
#> [210] "Katharine Sherratt"                                                             
#> [211] "Kazuki Yoshida"                                                                 
#> [212] "Kentaro Hirose"                                                                 
#> [213] "Kevin Spiegel"                                                                  
#> [214] "Kevin Weiss"                                                                    
#> [215] "Kevis Weiss"                                                                    
#> [216] "Kiesha Prem"                                                                    
#> [217] "Klaus Schliep"                                                                  
#> [218] "Kosuke Imai"                                                                    
#> [219] "Kristjan Eldjarn"                                                               
#> [220] "L. Held"                                                                        
#> [221] "LSHTM"                                                                          
#> [222] "Laha Ale"                                                                       
#> [223] "Lampros Bouranis"                                                               
#> [224] "Lander Willem"                                                                  
#> [225] "Lang"                                                                           
#> [226] "Laura Espinosa"                                                                 
#> [227] "Lei Liu"                                                                        
#> [228] "Leonhard Held"                                                                  
#> [229] "Li Wang"                                                                        
#> [230] "Libor Morkovsky"                                                                
#> [231] "Liesbeth C. de Wreede"                                                          
#> [232] "Liesbeth de Wreede"                                                             
#> [233] "Lifang Hou"                                                                     
#> [234] "Linda R. Petzold"                                                               
#> [235] "Lindsay V. Clark"                                                               
#> [236] "Lloyd Chapman"                                                                  
#> [237] "London School of Hygiene"                                                       
#> [238] "Lopaka Lee"                                                                     
#> [239] "Lore Merdrignac"                                                                
#> [240] "Lozano Jose E."                                                                 
#> [241] "Luke Keele"                                                                     
#> [242] "M Rantanen"                                                                     
#> [243] "M. Hofmann"                                                                     
#> [244] "M. Höhle"                                                                       
#> [245] "M. Hü¾\u008d¦\u0098¼hle"                                                        
#> [246] "M. Paul"                                                                        
#> [247] "M. Salmon"                                                                      
#> [248] "M. Virtanen"                                                                    
#> [249] "Maelle Salmon"                                                                  
#> [250] "Marcelo Ponce"                                                                  
#> [251] "Margot Einoder-Moreno"                                                          
#> [252] "Maria Bekker-Nielsen Dunbar"                                                    
#> [253] "Maria Noremark"                                                                 
#> [254] "Maria Prieto Gonzalez"                                                          
#> [255] "Marie- Pauline Beugin"                                                          
#> [256] "Marie-Pauline Beugin"                                                           
#> [257] "Mark Clements"                                                                  
#> [258] "Mark Jit"                                                                       
#> [259] "Mark Myatt"                                                                     
#> [260] "Mark Stevenson"                                                                 
#> [261] "Marta Fiocco"                                                                   
#> [262] "Martina Morris"                                                                 
#> [263] "Martyn Plummer"                                                                 
#> [264] "Matheus Costa"                                                                  
#> [265] "Mathew Jay"                                                                     
#> [266] "Mathias Hofmann"                                                                
#> [267] "Matthew J. Ferrari"                                                             
#> [268] "Matthijs S. Berends"                                                            
#> [269] "Matti Rantanen"                                                                 
#> [270] "Mauricio van den Berg"                                                          
#> [271] "Maëlle Salmon"                                                                  
#> [272] "Medicine"                                                                       
#> [273] "Michael A Rotondi"                                                              
#> [274] "Michael DeWitt"                                                                 
#> [275] "Michael Hills"                                                                  
#> [276] "Michael Hoehle"                                                                 
#> [277] "Michael Höhle"                                                                  
#> [278] "Michael Lavine"                                                                 
#> [279] "Michael P. Fay"                                                                 
#> [280] "Michael P. Fay User"                                                            
#> [281] "Michaela Paul"                                                                  
#> [282] "Michelle Ross"                                                                  
#> [283] "Mikael Moise"                                                                   
#> [284] "Mikko Virtanen"                                                                 
#> [285] "Minh Trinh"                                                                     
#> [286] "Mohammed Ali"                                                                   
#> [287] "Morgane Vallee"                                                                 
#> [288] "Nagendra Singanallur"                                                           
#> [289] "Neil M. Ferguson"                                                               
#> [290] "Nianqiao Ju"                                                                    
#> [291] "Nicholas G. Reich"                                                              
#> [292] "Nick Golding"                                                                   
#> [293] "Niel Hens"                                                                      
#> [294] "Nikos Bosse"                                                                    
#> [295] "Nistara Randhawa"                                                               
#> [296] "Nuredin Mohammed"                                                               
#> [297] "Oliver Watson"                                                                  
#> [298] "Ottar N. Bjornstad"                                                             
#> [299] "P. Alex Demarsh"                                                                
#> [300] "Paola Sebastiani"                                                               
#> [301] "Paul"                                                                           
#> [302] "Paul Bastide"                                                                   
#> [303] "Paul Mee"                                                                       
#> [304] "Paula Moraga"                                                                   
#> [305] "Pavel Dimens"                                                                   
#> [306] "Pavol Bauer"                                                                    
#> [307] "Pawel Piatkowski"                                                               
#> [308] "Peter Dutey-Magni"                                                              
#> [309] "Peter Ellis"                                                                    
#> [310] "Peter Li"                                                                       
#> [311] "Peter N. Brown"                                                                 
#> [312] "Peter Solymos"                                                                  
#> [313] "Petra Bäumler"                                                                  
#> [314] "Philippe Lemey"                                                                 
#> [315] "Pierre Nouvellet"                                                               
#> [316] "Pierre-Yves Boelle"                                                             
#> [317] "Pietro Coletti"                                                                 
#> [318] "Pietro Monticone"                                                               
#> [319] "Pratik Gupte"                                                                   
#> [320] "Pratik R. Gupte"                                                                
#> [321] "Qingyuan Zhao"                                                                  
#> [322] "Qingzhao Yu"                                                                    
#> [323] "R Core Team"                                                                    
#> [324] "R. Woodrow Setzer"                                                              
#> [325] "RJ Ewing"                                                                       
#> [326] "Rami Krispin"                                                                   
#> [327] "Rebecca DeCamp"                                                                 
#> [328] "Reshma Kassanjee"                                                               
#> [329] "Rich FitzJohn"                                                                  
#> [330] "Rich Fitzjohn"                                                                  
#> [331] "Riebler"                                                                        
#> [332] "Rob Deardon"                                                                    
#> [333] "Robert Gentleman"                                                               
#> [334] "Robin Eriksson"                                                                 
#> [335] "Robin Thompson"                                                                 
#> [336] "Rogier P. Schade"                                                               
#> [337] "Rolina van Gaalen"                                                              
#> [338] "Roman Lustrik"                                                                  
#> [339] "Ron Thornton"                                                                   
#> [340] "Ronald B. Geskus"                                                               
#> [341] "Ronald Galiwango"                                                               
#> [342] "Ronald Geskus"                                                                  
#> [343] "Rosalind Eggo"                                                                  
#> [344] "Ross Lazarus"                                                                   
#> [345] "Ryan Kyle"                                                                      
#> [346] "S. C. Kou"                                                                      
#> [347] "S. Meyer"                                                                       
#> [348] "S. Steiner"                                                                     
#> [349] "Salil Goyal"                                                                    
#> [350] "Salla E. Toikkanen"                                                             
#> [351] "Sam Abbott"                                                                     
#> [352] "Sam Clifford"                                                                   
#> [353] "Samuel Jenness"                                                                 
#> [354] "Sang Woo Park"                                                                  
#> [355] "Sangeeta Bhatia"                                                                
#> [356] "Sarah Connor"                                                                   
#> [357] "Sarah Pirikahu"                                                                 
#> [358] "Sarah V Leavitt"                                                                
#> [359] "Sebastian Funk"                                                                 
#> [360] "Sebastian Lequime"                                                              
#> [361] "Sebastian Meyer"                                                                
#> [362] "Shaoqian Ma"                                                                    
#> [363] "Shaoyang Ning"                                                                  
#> [364] "Shawnee Anderson"                                                               
#> [365] "Shi Zhao"                                                                       
#> [366] "Shihao Yang"                                                                    
#> [367] "Shikun Li"                                                                      
#> [368] "Simon Cauchemez"                                                                
#> [369] "Simon Daniel"                                                                   
#> [370] "Simon Dellicour"                                                                
#> [371] "Simon Firestone"                                                                
#> [372] "Simon Frost"                                                                    
#> [373] "Simon N. Wood"                                                                  
#> [374] "Sina Solaimanpour"                                                              
#> [375] "Sinead E. Morris"                                                               
#> [376] "Skye Bender-deMoll"                                                             
#> [377] "Sofia Ny"                                                                       
#> [378] "Sophie Meakin"                                                                  
#> [379] "Spencer Hall"                                                                   
#> [380] "Stefan Engblom"                                                                 
#> [381] "Stefan Steiner"                                                                 
#> [382] "Stefan Widgren"                                                                 
#> [383] "Stefan Widgren Maria Noremark"                                                  
#> [384] "Stefano Ongarello"                                                              
#> [385] "Stephen P. Ellner"                                                              
#> [386] "Steve Ellner"                                                                   
#> [387] "Steven G. Johnson"                                                              
#> [388] "Steven M. Goodreau"                                                             
#> [389] "Steven P. Millard"                                                              
#> [390] "Suzanne Keddie"                                                                 
#> [391] "T. Correa"                                                                      
#> [392] "Technology"                                                                     
#> [393] "Telmo Nunes"                                                                    
#> [394] "Teppei Yamamoto"                                                                
#> [395] "Thais Correa"                                                                   
#> [396] "The R Core Team"                                                                
#> [397] "Thibaut Jombart"                                                                
#> [398] "Thijs Janzen"                                                                   
#> [399] "Thomas Crellen"                                                                 
#> [400] "Thomas Czernichow"                                                              
#> [401] "Thomas Mollet"                                                                  
#> [402] "Thomas Obadia"                                                                  
#> [403] "Thomas Petzoldt"                                                                
#> [404] "Thomas Rosendal"                                                                
#> [405] "Tiago Maié"                                                                     
#> [406] "Tim Lucas"                                                                      
#> [407] "Tim Russell"                                                                    
#> [408] "Tim Taylor"                                                                     
#> [409] "Tobias Erik Reiners"                                                            
#> [410] "Tomas Aragon"                                                                   
#> [411] "Tomas J. Aragon"                                                                
#> [412] "Tomas J. Aragon Developer"                                                      
#> [413] "Tommi Karki"                                                                    
#> [414] "Tropical Medicine"                                                              
#> [415] "V. Wimmer"                                                                      
#> [416] "V. Wimmer. A few code segments are modified versions of code by The R Core Team"
#> [417] "VP Nagraj"                                                                      
#> [418] "Valentin Wimmer"                                                                
#> [419] "Vineetha Warriyar K. V"                                                         
#> [420] "Vineetha Warriyar K. V."                                                        
#> [421] "Vineetha Warriyar. K. V"                                                        
#> [422] "Virasakdi Chongsuvivatwong"                                                     
#> [423] "Vladimir Mikryukov"                                                             
#> [424] "Waleed Almutiry"                                                                
#> [425] "Wei Wei"                                                                        
#> [426] "Weihuang Wong"                                                                  
#> [427] "Weiliang Qiu"                                                                   
#> [428] "Wes Hinsley"                                                                    
#> [429] "Xavier Didelot"                                                                 
#> [430] "Yinan Zheng"                                                                    
#> [431] "Yongyou Zhang"                                                                  
#> [432] "Youcef Saad"                                                                    
#> [433] "Yuanwei Xu"                                                                     
#> [434] "Yuruo Li"                                                                       
#> [435] "Yusuke Asai"                                                                    
#> [436] "Zhian N. Kamvar"                                                                
#> [437] "data.org"                                                                       
#> [438] "odepack authors"